+++ lib/zakuro/calculation/range/transfer/western_date_allocation.rb
@@ -39,7 +39,7 @@ module Zakuro
                 new_year_date: new_year_date, months: months
               )
 
-              years[index] = Calculation::Base::Year.new(
+              years[index] = Base::Year.new(
                 multi_gengou: year.multi_gengou, new_year_date: new_year_date,
                 months: months, total_days: year.total_days
               )
@@ -55,9 +55,9 @@ module Zakuro
           def self.update_first_day_within_all_months(new_year_date:, months:)
             date = new_year_date.clone
             months.each_with_index do |month, index|
-              updated_month = Calculation::Monthly::Month.new(
+              updated_month = Monthly::Month.new(
                 month_label: month.month_label,
-                first_day: Calculation::Monthly::FirstDay.new(
+                first_day: Monthly::FirstDay.new(
                   remainder: month.first_day.remainder,
                   western_date: date
                 ),
